+++ astropy/wcs/__init__.py
@@ -28,3 +28,11 @@ try:
 except ImportError:
     if not _ASTROPY_SETUP_:
         raise
+
+
+def get_include():
+    """
+    Get the path to astropy.wcs's C header files.
+    """
+    import os
+    return os.path.join(os.path.dirname(__file__), "include")